Apply attribute packed to IDT_entry
C compilers are allowed to pad struct entries out, meaning that entry 2
may not come directly after entry 1, but instead empty space may be
inbetween the two, for various reasons. gcc provides
__attribute__((packed)) to guard against this, which when applied to a
struct guarentees that the struct will not be compiled with any empty
space. This is necessary for structs like IDT_entry, where they map to
structure the hardware is going to expect in a very specefic setup with
no padding.