Insert hlt instruction into kmain loop
Right now, kmain does a busy loop while we wait for interrupts. It's
much more effecient to put a hlt instruction inside of the loop, which
pauses the CPU until the next interrupt is recieved. This still requires
a loop because execution will continue after the hlt instruction once we
return from an interrupt.