Back-up register values in keyboard interrupt handler
Currently, we don't back up any of the registers before calling the
actual interrupt routine. This will become a problem because the
interrupt routine is going to trash these values, and they won't be
restored by 'iretd'. The simple solution is to push all the segments and
then all the registers onto the stack, and then pop them right before we
return from the interrupt.